+++ lib/puma/detect.rb
@@ -12,4 +12,14 @@ module Puma
   def self.windows?
     IS_WINDOWS
   end
+
+  IS_TRUFFLE = RUBY_ENGINE == 'truffleruby'
+
+  def self.truffle?
+    IS_TRUFFLE
+  end
+
+  def self.mri?
+    RUBY_ENGINE == 'ruby' || RUBY_ENGINE.nil?
+  end
 end