+++ test/test_cli.rb
@@ -55,7 +55,8 @@ class TestCLI < Minitest::Test
     body = s.read
     s.close
 
-    assert_match(/{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":16,"max_threads":16,"requests_count":0}/, body.split(/\r?\n/).last)
+    dmt = Puma::Configuration.new.default_max_threads
+    assert_match(/{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":#{dmt},"max_threads":#{dmt},"requests_count":0}/, body.split(/\r?\n/).last)
 
   ensure
     cli.launcher.stop
@@ -89,7 +90,8 @@ class TestCLI < Minitest::Test
       body = http.request(req).body
     end
 
-    expected_stats = /{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":16,"max_threads":16,"requests_count":0}/
+    dmt = Puma::Configuration.new.default_max_threads
+    expected_stats = /{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":#{dmt},"max_threads":#{dmt}/
     assert_match(expected_stats, body.split(/\r?\n/).last)
 
   ensure
@@ -168,7 +170,8 @@ class TestCLI < Minitest::Test
     body = s.read
     s.close
 
-    assert_match(/{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":16,"max_threads":16,"requests_count":0}/, body.split("\r\n").last)
+    dmt = Puma::Configuration.new.default_max_threads
+    assert_match(/{"started_at":"\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z","backlog":0,"running":0,"pool_capacity":#{dmt},"max_threads":#{dmt},"requests_count":0}/, body.split("\r\n").last)
   ensure
     if UNIX_SKT_EXIST
       cli.launcher.stop